mhf_fno_plugin/requirements.txt
torch>=1.9.0
neuraloperator>=2.0.0
numpy